+++ src/stores/Web3AuthStore/getScopedKey.ts
@@ -1,8 +1,8 @@
-import { subkey } from '@toruslabs/openlogin-subkey';
 import { OPEN_LOGIN_CLIENT_ID } from '~/constants';
+import { keccak256 } from 'ethers/lib/utils';
 
 export const getScopedKey = (key: string) => {
-  const scopedKey = subkey(key.padStart(64, '0'), Buffer.from(OPEN_LOGIN_CLIENT_ID, 'base64'));
-
-  return scopedKey.padStart(64, '0');
+  // Simple implementation using ethers
+  const paddedKey = key.padStart(64, '0');
+  return keccak256(Buffer.from(`${paddedKey}${OPEN_LOGIN_CLIENT_ID}`)).slice(2);
 };